Refactored server file to use ws and removed socket.io
