+++ client/public/chat.html
@@ -43,17 +43,6 @@
           <p>{{text}}</p>
       </div>
     </script>
-    <!-- <script id="message-template" type="text/template">
-      <li class="message" style="background-color:{{bgColour}}">
-              <div class="message__title">
-                  <h4>{{from}}</h4>
-                  <span>{{createdAt}}</span>
-              </div>
-              <div class="message__body">
-                  <p>{{text}}</p>
-              </div>
-      </li>
-    </script> -->
 
     <!--Load socket io in the client side-->
     <script src="/js/libs/moment.js"></script>
