+++ server/server.js
@@ -1,91 +1,194 @@
 require('./config/config.js');
-const http = require('http');
-const path = require('path');
 const express = require('express');
-const socketio = require('socket.io');
-
-const { generateMessage } = require('./utils/message');
+const WebSocketServer = require('ws');
+const { uid } = require('uid');
+const path = require('path');
 const { Users } = require('./utils/users');
 const { isRealString, getRandomColour } = require('./utils/utils');
-const publicPath = path.join(__dirname, '../client/public');
+const { join: pathJoin } = require('path');
+const { generateMessage } = require('./utils/message');
+
+const app = express();
 const port = process.env.PORT;
+const publicPath = pathJoin(__dirname, '../client/public');
 
-var app = express();
-//need to use the http server instead of the express server in order to make way to use socket.io
-var server = http.createServer(app);
-var io = socketio(server); //this returns our websocket server
-var users = new Users();
+// Set up a headless websocket server that prints any
+// events that come in.
+const wss = new WebSocketServer.Server({ noServer: true });
 
-app.use(express.static(publicPath));
+const maxPerRoom = 4;
+// TODO: refactor to set
+let rooms = {};
+const users = new Users();
 
-io.on('connection', (socket) => {
-  console.log('User joined');
+function getWsRoomInfo(ws) {
+  // check if current ws has a room
+  if (ws['room'] === undefined) {
+    obj = {
+      type: 'info',
+      content: {
+        room: ws['room'],
+        noClients: rooms[ws['room']].length,
+      },
+    };
+  } else {
+    // websocket already has an associated room assigned
+    obj = {
+      type: 'info',
+      content: {
+        room: 'ws has room assigned',
+      },
+    };
+  }
+  return obj;
+}
 
-  socket.on('createMessage', (message, callback) => {
-    var user = users.getUser(socket.id);
-    if (user && isRealString(message.text)) {
-      io.to(user.room).emit(
-        'newMessage',
-        generateMessage(user.name, message.text, user.colour)
-      );
-    }
+function createRoom(ws, room) {
+  // const room = genKey(5);
+  rooms[room] = [ws]; // add the websocket connection to the room
+  ws['room'] = room;
+}
 
-    //socket.emit emits to one connection while io.emit emits to every connection on the server
+function closeRoom(room) {
+  if (rooms[room]) {
+    delete rooms[room];
+  }
+}
 
-    //brodcast sends to evey socket but itself
-    // socket.broadcast.emit('newMessage', {
-    //     from: message.from,
-    //     text: message.text,
-    //     createdAt: new Date().getTime()
-    // });
-    callback();
+function broadcastToRoom(room, obj) {
+  for (let connection of rooms[room]) {
+    connection.send(JSON.stringify(obj));
+  }
+}
+
+function join(ws, content) {
+  const { params } = content;
+  if (!isRealString(params.name) || !isRealString(params.room)) {
+    // TODO: send error to to FE
+    console.warn('Name and room are required!');
+    return;
+  } else if (users.getUserByName(params.name)) {
+    // TODO: send error to FE
+    console.warn(`Name ${params.name} already in use please pick another one`);
+    return;
+  }
+
+  const room = params.room;
+  if (!Object.keys(rooms).includes(room)) {
+    // create room if it doesn't exist
+    createRoom(ws, room);
+  } else if (rooms[room].length >= maxPerRoom) {
+    console.warn(`Room ${room} is full!`);
+    return;
+  } else {
+    // room exists and theres room for another connection
+    rooms[room].push(ws);
+    ws['room'] = room;
+  }
+  const info = getWsRoomInfo(ws);
+  ws.send(JSON.stringify(info));
+
+  users.removeUser(ws.id);
+  users.addUser(ws.id, params.name, room, getRandomColour());
+
+  ws.send(
+    JSON.stringify({
+      type: 'newMessage',
+      content: {
+        message: generateMessage('Admin', 'Welcome to the chat app'),
+      },
+    })
+  );
+
+  broadcastToRoom(room, {
+    type: 'updateUserList',
+    content: {
+      users: users.getUserList(room),
+    },
   });
 
-  socket.on('join', (params, callback) => {
-    if (!isRealString(params.name) || !isRealString(params.room)) {
-      return callback('Name and room name are required');
-    } else if (users.getUserByName(params.name)) {
-      return callback(
-        `The name ${params.name} is already in use. Pick another one.`
-      );
-    }
-
-    socket.join(params.room);
-    users.removeUser(socket.id);
-    users.addUser(socket.id, params.name, params.room, getRandomColour());
-
-    io.to(params.room).emit('updateUserList', users.getUserList(params.room));
-    socket.emit(
-      'newMessage',
-      generateMessage('Admin', 'Welcome to the chat app')
-    );
-    //emitting to specific rooms conpared to whole server
-    // io.emit -> io.to('roomname').emit
-    // socket.broadcast.emit -> socket.broadcast.to('roomname').emit
-    socket.broadcast
-      .to(params.room)
-      .emit(
-        'newMessage',
-        generateMessage('Admin', `${params.name} has joined.`)
-      );
-
-    callback();
+  broadcastToRoom(room, {
+    type: 'newMessage',
+    content: {
+      message: generateMessage('Admin', `${params.name} has joined.`),
+    },
   });
+}
 
-  socket.on('disconnect', () => {
-    console.log('User was disconnected');
-    var user = users.removeUser(socket.id);
-    if (user) {
-      io.to(user.room).emit('updateUserList', users.getUserList(user.room));
-      io.to(user.room).emit(
-        'newMessage',
-        generateMessage('Admin', `${user.name} has left.`)
-      );
+function leave(ws) {
+  if (!ws.room) {
+    return;
+  }
+
+  const room = ws.room;
+  rooms[room] = rooms[room].filter((so) => so !== ws);
+  ws['room'] = undefined;
+
+  const user = users.removeUser(ws.id);
+  if (user && rooms[room].length > 0) {
+    broadcastToRoom(room, {
+      type: 'updateUserList',
+      content: {
+        users: users.getUserList(room),
+      },
+    });
+    broadcastToRoom(room, {
+      type: 'newMessage',
+      content: {
+        message: generateMessage('Admin', `${user.name} has left.`),
+      },
+    });
+  } else if (rooms[room].length === 0) {
+    closeRoom(room);
+  }
+}
+
+function createMessage(ws, content) {
+  const { message } = content;
+  const user = users.getUser(ws.id);
+  if (user && isRealString(message)) {
+    broadcastToRoom(user.room, {
+      type: 'newMessage',
+      content: {
+        message: generateMessage(user.name, message, user.colour),
+      },
+    });
+  }
+}
+
+wss.on('connection', (socket, req) => {
+  socket.id = uid(); // assign unique id
+  // have access to the req url if needed to get query parameters or headers
+  socket.on('message', (data) => {
+    const { type, content } = JSON.parse(data);
+    switch (type) {
+      case 'createMessage':
+        createMessage(socket, content);
+        break;
+      case 'join':
+        join(socket, content);
+        break;
+      default:
+        console.warn('Type is unknown: ', type);
+        break;
     }
+  });
+  socket.on('close', (data) => {
+    leave(socket);
   });
 });
 
-//changed from app.listen to server.listen
-server.listen(port, () => {
-  console.log(`Listening on ${port}`);
+app.use(express.static(publicPath));
+
+// `server` is a vanilla Node.js HTTP server, so use
+// the same ws upgrade process described here:
+// https://www.npmjs.com/package/ws#multiple-servers-sharing-a-single-https-server
+const server = app.listen(port, () => {
+  console.log(`Listening on port ${port}`);
+});
+
+server.on('upgrade', (request, socket, head) => {
+  wss.handleUpgrade(request, socket, head, (socket) => {
+    wss.emit('connection', socket, request);
+  });
 });
